addressed review comments for unit test
